create system/hardware/pc/hardware.h
