+++ release/files_common
@@ -216,6 +216,7 @@ system/hardware/tici/amplifier.py
 system/hardware/tici/updater
 system/hardware/tici/iwlist.py
 system/hardware/pc/__init__.py
+system/hardware/pc/hardware.h
 system/hardware/pc/hardware.py
 
 selfdrive/locationd/__init__.py
