+++ system/hardware/pc/hardware.h
@@ -1,0 +1,14 @@
+#pragma once
+
+#include "common/util.h"
+#include "system/hardware/base.h"
+
+class HardwarePC : public HardwareNone {
+public:
+  static std::string get_os_version() { return "openpilot for PC"; }
+  static std::string get_name() { return "pc"; };
+  static cereal::InitData::DeviceType get_device_type() { return cereal::InitData::DeviceType::PC; };
+  static bool PC() { return true; }
+  static bool TICI() { return util::getenv("TICI", 0) == 1; }
+  static bool AGNOS() { return util::getenv("TICI", 0) == 1; }
+};
